windows/requirements.txt
cffi==1.16.0
clr-loader==0.2.6
pycparser==2.21
pydantic==1.10.7
python-dotenv==1.0.0
pythonnet==3.0.3
pywin32==306
typing_extensions==4.8.0
